requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
optuna>=3.5
shap>=0.44
torch>=2.2
pyyaml>=6.0
joblib>=1.3
tqdm>=4.66
